Created Signature highlight groups to be used by themes
Created highlight groups SignatureMarkText, SignatureMarkLine, SignatureMarkerText and SignatureMarkerLine
This should hopefully simplify the colorscheme code and allow it to remain blissfully ignorant of Signature code

Additionally, creation of these highlight groups allows me to use it as a fallback option in the following scenario
Let's say g:SignatureMarkTextHLDynamic=1 and we're on a line that doesn't have any changes, so no signs are placed by
GitGutter or Signify. In this case, Signature needs a fallback option and that's where these highlight groups come in.
By default, these are linked to "Exception" and "WarningMsg" to preserve the traditional behavior.
If required, colorschemes can link it to something else and thereby have the signs fit into their scheme.