Fix for #72: Updated s:ForceRemoveGlobal to use :wviminfo!/:wshada!
This is because editing the viminfo/shada files no longer works as once
the buffer is saved, the information about the marks gets written right
back.

The default value of `g:ForceRemoveGlobalMark` has been changed to 0
since this is a very heavy hammer to use and goes against *vim's
deliberate design choice.